requirements.txt
# requirements.txt

streamlit
langchain
chromadb
sentence-transformers
google-generativeai
python-dotenv
coqui-tts
pydub
soundfile
